feat(service): new `POST /bases` route
