+++ src/main.rs
@@ -56,7 +56,7 @@ async fn main() -> std::io::Result<()> {
             .service(
                 web::scope("/api")
                 .service(service::get_units)
-                .service(service::post_create_base)
+                .service(service::bases::post)
             )
     })
     .bind(("127.0.0.1", 8080))?
